requirements.txt
langchain-mcp-adapters
langgraph
langchain-openai
langchain-groq
beautifulsoup4>=4.12.0
requests>=2.31.0
mcp>=0.1.0
python-dotenv>=1.0.0